test: write detailed tests with the new files
